Home health data fixed

+++ lib/model/home_health_data.dart
@@ -1,5 +1,5 @@
 class HomeHealthModel {
-  final int health;
+  final double health;
   final int totalTasks;
   final int completedTasks;
